revno: 15385 message:   minor fix, unnecessary unboxing
committer: Morten Olav Hansen <mortenoh@gmail.com>
branch nick: dhis2
timestamp: Fri 2014-05-23 17:31:37 +0200

+++ dhis-2/dhis-api/src/main/java/org/hisp/dhis/program/ProgramInstance.java
@@ -310,7 +310,7 @@ public class ProgramInstance
     @JacksonXmlProperty( namespace = DxfNamespaces.DXF_2_0 )
     public int getStatus()
     {
-        return status.intValue();
+        return status;
     }
 
     public void setStatus( Integer status )
